charity_api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "charity-api"
version = "1.0.0"
description = "Comprehensive charity evaluation API using multiple data sources"
authors = [{name = "Charity Analysis Team"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "pandas>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["charity_api"]